refactor: Add Create and Deploy modules for application management
